pyproject.toml
[project]
name = "tech-professional-coacher"
version = "0.1.0"
description = "Implementation of multi-agent achitecture to make an tech professional coacher application"
readme = "README.md"
authors = [
    { name = "lucasvittal2", email = "lucasvittal@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "pre-commit>=4.2.0",
    "dotenv>=0.9.9",
    "opentelemetry-api>=1.35.0",
    "opentelemetry-distro[otlp]>=0.56b0",
    "opentelemetry-exporter-otlp>=1.35.0",
    "opentelemetry-instrumentation-asyncio>=0.56b0",
    "opentelemetry-instrumentation-fastapi>=0.56b0",
    "opentelemetry-instrumentation-requests>=0.56b0",
    "opentelemetry-sdk>=1.35.0",
    "gunicorn>=23.0.0",
    "uvicorn>=0.35.0",
    "pyjwt[crypto]>=2.10.1",
    "cryptography>=45.0.5",
    "pip>=25.1.1",
    "pydantic-settings>=2.10.1",
    "opentelemetry-instrumentation>=0.56b0",
    "langsmith>=0.4.8",
    "openai>=1.97.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tech_professional_coacher"]
